refactor: service 분리 및 useHealth 수정 (#84)
