+++ src/features/health/ui/HealthPage.tsx
@@ -1,7 +1,5 @@
 'use client'
 
-import { useParams } from 'next/navigation'
-
 import { GRAPHTYPE } from '@/constants'
 import { DashBoardTemplate } from '@/features'
 import { useToggle } from '@/hooks'
@@ -12,9 +10,8 @@ import { ReportSection } from './ReportSection'
 import { WeightSection } from './WeightSection'
 
 export const HealthPage = () => {
-  const { userId } = useParams<{ userId: string }>()
   const [graphType, toggleGraphType] = useToggle(true)
-  const { data: healthPageData } = useFetchHealth({ userId, period: GRAPHTYPE[`${graphType}`].en })
+  const { data: healthPageData } = useFetchHealth(GRAPHTYPE[`${graphType}`].en)
 
   if (!healthPageData) {
     return (
