+++ src/hooks/apis/useHealth.ts
@@ -1,14 +1,18 @@
+'use client'
+
+import { useParams } from 'next/navigation'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 
 import { getHealth, postHealth, putHealth } from '@/apis/health'
-import type { GetPhysicalRequest, PutPhysicalRequest } from '@/types'
+import type { PutPhysicalRequest } from '@/types'
 
 const queryKeys = {
   all: ['health'] as const,
   page: (userId: string, period: string) => [...queryKeys.all, userId, period],
 }
 
-export const useFetchHealth = ({ userId, period }: GetPhysicalRequest) => {
+export const useFetchHealth = (period: string) => {
+  const { userId } = useParams<{ userId: string }>()
   return useQuery({
     queryKey: queryKeys.page(userId, period),
     queryFn: () => getHealth({ userId, period }),
